+++ CHANGES.rst
@@ -36,7 +36,7 @@ CHANGES
 
 - Revert `resp.url` back to `str`, introduce `resp.url_obj` #1292
 
-- Raise ValueError if BasicAuth login has a ":" character #1307
+- Quote login and password for basic auth separately #1307
 
 -